Still implementing voting. klappfix for uservote implemented
