+++ frontend/src/components/PlanItem/PlanItemDisplayDetails.tsx
@@ -1,6 +1,6 @@
 import {PlanItem} from "../../model/PlanItem";
 import React from "react";
-import DisplayVoting from "../Voting/DisplayVoting";
+import VotingScorecard from "../Voting/VotingScorecard";
 
 type PlanItemDisplayProps = {
     detailedPlanItem: PlanItem
@@ -17,7 +17,7 @@ export default function PlanItemDisplayDetails({detailedPlanItem}: PlanItemDispl
                 <div>Beschreibung: {detailedPlanItem.planDescription}</div>
                 <div>Geplant am: {detailedPlanItem.plannedOn}</div>
                 <div>Planer: {detailedPlanItem.plannedBy}</div>
-                <div><DisplayVoting detailedPlanItem={detailedPlanItem}/></div>
+                <div><VotingScorecard detailedPlanItem={detailedPlanItem}/></div>
                 <div>Teilnehmer: {detailedPlanItem.finalGang.filter(item => (!item.isPlanned)).map(item => <div
                     key={item.id}>{item.username}</div>)}</div>
                 <div>Optionen: {detailedPlanItem.dateOptions.map(item => <div
